Sanitize term field before querying for existing position type term
There is an oddity in the way term names are stored that forces them to
be sanitized using kses before being stored. This includes encoding
ampersand characters to `&amp;`, so that they're stored in an encoded
fashion in the db. However, `get_term_by()` was not also doing the same
encoding, with the result that the sql query in the latter function was
failing to match anything.

I think this is a WP bug. See http://core.trac.wordpress.org/ticket/11311#comment:14

My workaround is to sanitize the search value before passing to get_term_by()

See CAC #2728