Better IE compat
We want to force IE not to use compatibility mode (as it's set to
do at the Graduate Center). So we send the appropriate <meta> tag.

This was not working with the fix introduced in 05a308f7df8865908523b447289290bc360bf2a9
for two reasons. One, cacap templates are not technically part of
the bp-nelo theme. Two, there is a problem with the X-UA-Compatible meta
tag and transitional doctypes (so we switch to html)

See #2798